# core/api/rest/openapi.yaml
...
openapi: 3.1.0
info:
  title: CypherPanel API
  description: >-
    Operator-facing control-plane API covering Phases 1–3: authentication and
    teams/roles, the server lifecycle, projects/environments, applications and
    the deploy pipeline, managed databases with S3 backups/restore, preview
    environments, notifications, and scheduled tasks. Authentication is a bearer
    session token from /api/v1/auth/login; the GitHub webhook authenticates by
    per-application HMAC instead. Every project-scoped route is authorized by
    team membership (a non-member sees 404, insufficient rank 403).


    Every response — success, error and SSE stream alike — carries an
    `X-Request-Id` header (`components/headers/RequestId`), and every JSON
    error body repeats it as `trace_id`. It is the value to quote in a bug
    report and the key to search for in `GET /api/v1/panel/logs`. Individual
    responses reference the header only where a generated client benefits;
    it is present on all of them.
  version: 0.3.0
servers:
  - url: /
...
paths:
...
  /api/v1/projects/{id}/environments:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Project ID (prj_… prefix).
    get:
      operationId: listEnvironments
      tags: [projects]
      summary: List a project's environments
      responses:
        "200":
          description: The project's environments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Environment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createEnvironment
      tags: [projects]
      summary: Add an environment to a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvironmentRequest"
      responses:
        "201":
          description: Environment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/environments/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string }, description: Environment ID. }
    patch:
      operationId: patchEnvironment
      tags: [projects]
      summary: Rename a standing environment (team admin)
      description: >
        A preview environment belongs to its pull request and is refused —
        renaming one by hand desynchronises it from the PR that made it.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PatchEnvironmentRequest" } } }
      responses:
        "200": { description: Renamed., content: { application/json: { schema: { $ref: "#/components/schemas/Environment" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: "A preview, or the name is taken in the project.", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    delete:
      operationId: deleteEnvironment
      tags: [projects]
      summary: Remove a standing environment (team admin)
      description: >
        Refused for a preview (its pull request owns it), for the last standing
        environment a project has, and while resources remain inside — the same
        protection deleting a project gets.
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: "A preview, the last environment, or resources remain.", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Session token from /api/v1/auth/login.
  headers:
    RequestId:
      description: >-
        Correlation id for this request, minted by the API's outermost
        middleware and present on every response. It is repeated as `trace_id`
        in every JSON error body, and it is the key to search for in
        `GET /api/v1/panel/logs`. An inbound `X-Request-Id` is honoured only
        when the peer is a configured trusted proxy and the value is a short
        printable token; otherwise a fresh id is minted.
      schema: { type: string }
      example: trace_9f3a-11bd-4c02-7e51
  responses:
    BadRequest:
      description: Malformed or invalid request body.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired session token.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated but the caller's role is insufficient for this action.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: >-
        No such resource — also returned when the caller is not a member of the
        owning team, so resource existence never leaks across tenants.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message; never contains secrets.
        trace_id:
          type: string
          description: >-
            The value of this response's `X-Request-Id` header, repeated in the
            body so a screenshot of a failure carries it. Paste it into a bug
            report or `GET /api/v1/panel/logs` to find the request. Present on
            every response served through the API's middleware.
          example: trace_9f3a-11bd-4c02-7e51
        retry_after_seconds:
          type: integer
          minimum: 1
          description: >-
            Present only on `429`: whole seconds until the throttle lifts, the
            same value as the `Retry-After` header. Count down against it rather
            than guessing.
...
    PatchEnvironmentRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
...
    Environment:
      type: object
      required: [id, project_id, name, created_at]
      properties:
        id:
          type: string
          description: env_… prefixed ID.
        project_id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        kind:
          type: string
          enum: [production, standard, preview]
          description: >-
            A preview is created and destroyed by its pull request and cannot be
            renamed or deleted by hand.
        is_default: { type: boolean }
...
    CreateEnvironmentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
...
security:
  - bearer: []